Rename rule module to harmonization_rule
- Rename rule.py to harmonization_rule.py
- Update imports and exports
- Keep Rule alias for compatibility

Closes #60
