+++ src/harmonization_framework/__init__.py
@@ -1,1 +1,2 @@
 from .harmonize import harmonize_dataset, harmonize_file
+from .harmonization_rule import HarmonizationRule
